backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
pydantic==2.11.9
pydantic-settings==2.11.0
email-validator==2.3.0
sqlalchemy==2.0.43
alembic==1.16.5
psycopg2-binary==2.9.10
cloud-sql-python-connector==1.18.4
pg8000==1.31.5
redis==6.4.0
google-cloud-dialogflow-cx>=1.30.0
google-cloud-aiplatform==1.118.0
google-cloud-resource-manager==1.14.2
google-generativeai==0.8.5
google-cloud-storage<3.0.0,>=2.10.0
google-auth>=2.28.0
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pandas==2.3.3
openpyxl==3.1.5
pytest==8.4.2
pytest-asyncio==1.2.0
httpx==0.28.1
python-dotenv==1.1.1
aiofiles==24.1.0
websockets==15.0.1
chardet==5.2.0
beautifulsoup4==4.14.2
